+++ lib/src/application/themes/default_task_theme.dart
@@ -1,5 +1,4 @@
 import 'package:commander_ui/commander_ui.dart';
-import 'package:commander_ui/src/domains/themes/task_theme.dart';
 
 final class DefaultTaskTheme implements TaskTheme {
   @override